cleaned code a bit and fixed formatting/imports/etc
