+++ pdr_backend/lake/test/test_gql_data_factory.py
@@ -4,10 +4,9 @@ import sys
 from pdr_backend.ppss.ppss import mock_ppss
 from pdr_backend.lake.gql_data_factory import GQLDataFactory
 from pdr_backend.util.time_types import UnixTimeMs
+from pdr_backend.lake.table import TableType, get_table_name
 from pdr_backend.lake.table_registry import TableRegistry
 from pdr_backend.lake.test.resources import _clean_up_table_registry
-from pdr_backend.lake.persistent_data_store import PersistentDataStore
-from pdr_backend.lake.plutil import get_table_name, TableType
 
 
 def test_gql_data_factory():
@@ -68,15 +67,14 @@ def test_update_end_to_end(_mock_fetch_gql, tmpdir):
     tables = TableRegistry().get_tables().items()
     assert count_updates == len(tables)
 
+
 def test_update_partial_then_resume(
-    _mock_fetch_gql,
-    _get_test_PDS,
-    _clean_up_test_folder,
-    tmpdir):
+    _mock_fetch_gql, _get_test_PDS, _clean_up_test_folder, tmpdir
+):
     """
     Test GQLDataFactory should update end-to-end, but fail in the middle
     Work 1: Update csv data (11-03 -> 11-05) and then fail inserting to db
-    Work 2: Then, update and verify csv has new records (11-05 -> 11-07) + db table has all records (11-03 -> 11-07)
+    Work 2: Update and verify new records (11-05 -> 11-07) + table has all records (11-03 -> 11-07)
     """
     _clean_up_test_folder(tmpdir)
     _clean_up_table_registry()
@@ -90,7 +88,7 @@ def test_update_partial_then_resume(
         st_timestr=st_timestr,
         fin_timestr=fin_timestr,
     )
-    
+
     # Work 1: update csv files and insert into temp table
     fns = {
         "pdr_predictions": _mock_fetch_gql,
@@ -100,26 +98,28 @@ def test_update_partial_then_resume(
     }
 
     gql_data_factory = GQLDataFactory(ppss)
-    with patch("pdr_backend.lake.gql_data_factory.GQLDataFactory._move_from_temp_tables_to_live") as mock_move_from_temp_tables_to_live:
+    with patch(
+        "pdr_backend.lake.gql_data_factory.GQLDataFactory._move_from_temp_tables_to_live"
+    ):
         for table_name in gql_data_factory.record_config["fetch_functions"]:
-            gql_data_factory.record_config["fetch_functions"][table_name] = fns[table_name]    
+            gql_data_factory.record_config["fetch_functions"][table_name] = fns[
+                table_name
+            ]
         gql_data_factory._update()
 
         # Verify records exist in temp pred tables
         pds = _get_test_PDS(ppss.lake_ss.lake_dir)
         target_table_name = get_table_name("pdr_predictions", TableType.TEMP)
-        target_records = pds.query_data(
-            "SELECT * FROM {}".format(target_table_name)
-        )
+        target_records = pds.query_data("SELECT * FROM {}".format(target_table_name))
 
         assert len(target_records) == 2
         assert target_records["pair"].to_list() == ["ADA/USDT", "BNB/USDT"]
         assert target_records["timestamp"].to_list() == [1699038000000, 1699124400000]
-    
+
     # Work 2: apply simulated error, update ppss "poorly", and verify it works as expected
     # Inject error by dropping db tables
     for table_name in gql_data_factory.record_config["fetch_functions"]:
-        pds.drop_table(get_table_name(table_name, TableType.TEMP))   
+        pds.drop_table(get_table_name(table_name, TableType.TEMP))
 
     # manipulate ppss poorly and run gql_data_factory again
     gql_data_factory.ppss.lake_ss.d["st_timestr"] = "2023-11-01"
@@ -128,12 +128,20 @@ def test_update_partial_then_resume(
 
     # Verify expected records were written to db
     target_table_name = get_table_name("pdr_predictions")
-    target_records = pds.query_data(
-        "SELECT * FROM {}".format(target_table_name)
-    )
+    target_records = pds.query_data("SELECT * FROM {}".format(target_table_name))
     assert len(target_records) == 4
-    assert target_records["pair"].to_list() == ["ADA/USDT", "BNB/USDT", "ETH/USDT", "ETH/USDT"]
-    assert target_records["timestamp"].to_list() == [1699038000000, 1699124400000, 1699214400000, 1699300800000]
+    assert target_records["pair"].to_list() == [
+        "ADA/USDT",
+        "BNB/USDT",
+        "ETH/USDT",
+        "ETH/USDT",
+    ]
+    assert target_records["timestamp"].to_list() == [
+        1699038000000,
+        1699124400000,
+        1699214400000,
+        1699300800000,
+    ]
 
 
 @patch("pdr_backend.lake.gql_data_factory.GQLDataFactory._update")
