+++ pdr_backend/lake/test/test_table_bronze_pdr_predictions.py
@@ -10,13 +10,12 @@ from pdr_backend.lake.table_pdr_predictions import (
     predictions_schema,
     predictions_table_name,
 )
-from pdr_backend.lake.table import Table
+from pdr_backend.lake.table import Table, get_table_name
 from pdr_backend.lake.table_pdr_truevals import truevals_schema, truevals_table_name
 from pdr_backend.lake.table_pdr_payouts import payouts_schema, payouts_table_name
 from pdr_backend.lake.persistent_data_store import PersistentDataStore
 from pdr_backend.lake.test.resources import _clean_up_table_registry
 from pdr_backend.util.time_types import UnixTimeMs
-from pdr_backend.lake.plutil import get_table_name, TableType
 
 
 @enforce_types
@@ -49,15 +48,9 @@ def test_table_bronze_pdr_predictions(
     }
 
     # Work 1: Append all data onto bronze_table
-    gql_tables["pdr_predictions"].append_to_storage(
-        _gql_datafactory_etl_predictions_df
-    )
-    gql_tables["pdr_truevals"].append_to_storage(
-        _gql_datafactory_etl_truevals_df
-    )
-    gql_tables["pdr_payouts"].append_to_storage(
-        _gql_datafactory_etl_payouts_df
-    )
+    gql_tables["pdr_predictions"].append_to_storage(_gql_datafactory_etl_predictions_df)
+    gql_tables["pdr_truevals"].append_to_storage(_gql_datafactory_etl_truevals_df)
+    gql_tables["pdr_payouts"].append_to_storage(_gql_datafactory_etl_payouts_df)
 
     pds = PersistentDataStore(ppss.lake_ss.lake_dir)
     # truevals should have 6
